+++ src/Validation.Helper/ConfigurationException.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Runtime.Serialization;
+
+namespace NuGet.Jobs.Validation.Helper
+{
+    [Serializable]
+    internal class ConfigurationException : Exception
+    {
+        public ConfigurationException()
+        {
+        }
+
+        public ConfigurationException(string message) : base(message)
+        {
+        }
+
+        public ConfigurationException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        protected ConfigurationException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}
